Initial Commit.
I made this last year, but thought I might as well make it public and track how I refine my code. It is a near-perfect space invaders clone (as far as what I've seen online and played in arcades goes), but the code is large and messy. A lot of the enhancements I make here are probably going to apply to a larger project I'm working on.